+++ backend/src/main/java/com/luiswiederhold/backend/flashcards/FlashcardService.java
@@ -1,6 +1,6 @@
 package com.luiswiederhold.backend.flashcards;
 
-import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
+import com.luiswiederhold.backend.DTO.FlashcardSideDTO;
 import com.luiswiederhold.backend.exception.LowConfidenceException;
 import com.luiswiederhold.backend.flashcards.imagestorage.ImageStorageService;
 import org.springframework.beans.factory.annotation.Autowired;
