+++ backend/src/main/java/com/luiswiederhold/backend/flashcards/imagestorage/DiskImageStorageService.java
@@ -1,6 +1,6 @@
 package com.luiswiederhold.backend.flashcards.imagestorage;
 
-import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
+import com.luiswiederhold.backend.DTO.FlashcardSideDTO;
 import com.luiswiederhold.backend.flashcards.Flashcard;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
