+++ backend/src/main/java/com/luiswiederhold/backend/flashcards/imagestorage/S3ImageStorageService.java
@@ -1,6 +1,6 @@
 package com.luiswiederhold.backend.flashcards.imagestorage;
 
-import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
+import com.luiswiederhold.backend.DTO.FlashcardSideDTO;
 import com.luiswiederhold.backend.flashcards.Flashcard;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
