+++ backend/src/test/java/com/luiswiederhold/backend/flashcards/S3ImageStorageServiceTest.java
@@ -1,6 +1,6 @@
 package com.luiswiederhold.backend.flashcards;
 
-import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
+import com.luiswiederhold.backend.DTO.FlashcardSideDTO;
 import com.luiswiederhold.backend.flashcards.imagestorage.S3ImageStorageService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
